server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aerich>=0.8.0",
    "bcrypt>=4.2.1",
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.6",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.10.1",
    "tortoise-orm>=0.23.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "aenum>=3.1.15",
]
